refactor to put down code to concrete classes
