+++ src/main/java/org/jenkinsci/plugins/github_branch_source/ForkPullRequestDiscoveryTrait.java
@@ -181,7 +181,7 @@ public class ForkPullRequestDiscoveryTrait extends SCMSourceTrait {
          */
         @Override
         public Class<? extends SCMSource> getSourceClass() {
-            return GitHubSCMSourceAbstract.class;
+            return AbstractGitHubSCMSource.class;
         }
 
         /**
